backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.7,<4.8
psycopg[binary]>=3.2,<3.3
